# src/frontend/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "FlowDeck.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost/"
    }
  ],
  "paths": {
    "/api/workflows/{definitionId}/instances": {
      "post": {
        "tags": [
          "WorkflowEndpoints"
        ],
        "summary": "Starts a new instance of a workflow definition.",
        "operationId": "StartWorkflowInstance",
        "parameters": [
          {
            "name": "definitionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "version",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartInstanceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/workflows": {
      "get": {
        "tags": [
          "WorkflowEndpoints"
        ],
        "summary": "Lists the workflow definitions this host has registered.",
        "operationId": "ListWorkflowDefinitions",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkflowDefinitionResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/instances/{instanceId}": {
      "get": {
        "tags": [
          "InstanceEndpoints"
        ],
        "summary": "Retrieves a single workflow instance.",
        "operationId": "GetWorkflowInstance",
        "parameters": [
          {
            "name": "instanceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstanceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/instances": {
      "get": {
        "tags": [
          "InstanceEndpoints"
        ],
        "summary": "Lists workflow instances, newest first.",
        "operationId": "ListWorkflowInstances",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/InstanceStatus"
            }
          },
          {
            "name": "definitionId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstancePage"
                }
              }
            }
          }
        }
      }
    },
    "/api/instances/{instanceId}/cancel": {
      "post": {
        "tags": [
          "InstanceEndpoints"
        ],
        "summary": "Stops a workflow instance permanently.",
        "operationId": "CancelWorkflowInstance",
        "parameters": [
          {
            "name": "instanceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstanceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/instances/{instanceId}/history": {
      "get": {
        "tags": [
          "InstanceEndpoints"
        ],
        "summary": "Reads an instance\u0027s execution history, in order.",
        "operationId": "GetWorkflowInstanceHistory",
        "parameters": [
          {
            "name": "instanceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StepHistoryResponse"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "InstancePage": {
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InstanceResponse"
            }
          },
          "total": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "InstanceResponse": {
        "required": [
          "id",
          "definitionId",
          "definitionVersion",
          "status",
          "currentStepIndex",
          "currentStepName",
          "createdAt",
          "completedAt",
          "failedStepName",
          "errorType",
          "errorMessage"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "definitionId": {
            "type": "string"
          },
          "definitionVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "status": {
            "$ref": "#/components/schemas/InstanceStatus"
          },
          "currentStepIndex": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "currentStepName": {
            "type": [
              "null",
              "string"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "failedStepName": {
            "type": [
              "null",
              "string"
            ]
          },
          "errorType": {
            "type": [
              "null",
              "string"
            ]
          },
          "errorMessage": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "InstanceStatus": {
        "enum": [
          "Running",
          "Suspended",
          "Completed",
          "Failed",
          "Cancelled"
        ]
      },
      "StartInstanceResponse": {
        "required": [
          "instanceId",
          "status"
        ],
        "type": "object",
        "properties": {
          "instanceId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "$ref": "#/components/schemas/InstanceStatus"
          }
        }
      },
      "StepHistoryResponse": {
        "required": [
          "sequence",
          "stepName",
          "startedAt",
          "completedAt",
          "durationMs",
          "status",
          "errorType",
          "errorMessage"
        ],
        "type": "object",
        "properties": {
          "sequence": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "stepName": {
            "type": "string"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time"
          },
          "durationMs": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d\u002B)?(?:[eE][\u002B-]?\\d\u002B)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "status": {
            "$ref": "#/components/schemas/StepStatus"
          },
          "errorType": {
            "type": [
              "null",
              "string"
            ]
          },
          "errorMessage": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "StepStatus": {
        "enum": [
          "Success",
          "Failed"
        ]
      },
      "WorkflowDefinitionResponse": {
        "required": [
          "id",
          "version",
          "inputTypeName"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "version": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "inputTypeName": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "WorkflowEndpoints"
    },
    {
      "name": "InstanceEndpoints"
    }
  ]
}